Rename package name to paymentbuttons
