+++ PaymentButtons/src/main/java/com/paypal/android/paymentbuttons/PaymentButtonSize.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.paymentbutton
+package com.paypal.android.paymentbuttons
 
 import com.paypal.android.checkout.paymentbutton.error.createFormattedIllegalArgumentException
 import com.paypal.android.ui.R
