+++ docs/PayPalUI/README.md
@@ -25,7 +25,7 @@ The PayPalUI module allows you to render three buttons that can offer a set of c
 Add a `PayPalButton` to your layout XML:
 
 ```xml
-<com.paypal.android.ui.paymentbutton.PayPalButton
+<com.paypal.android.paymentbutton.PayPalButton
     android:id="@+id/paypal_button"
     android:layout_width="match_parent"
     android:layout_height="wrap_content" />
